requirements.txt
Flask==2.3.2
Flask-JWT-Extended==4.5.2
Flask-Limiter==2.6.2
Flask-Cors==4.0.1
Flask-Marshmallow==0.14.0
marshmallow==3.19.0
python-dotenv==0.21.0
requests==2.31.0
beautifulsoup4==4.12.2
google-generativeai==0.3.1
scikit-learn==1.3.0
numpy==1.24.3
gunicorn==20.1.0
lxml==4.9.3